2.0-search-agent/pyproject.toml
[project]
name = "2-0-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-openai>=1.2.1",
    "langchain-tavily>=0.2.18",
    "python-dotenv>=1.2.2",
]
